Big refactor.
