+++ wasm-runtime/polkadot/src/timestamp.rs
@@ -1,0 +1,9 @@
+use primitives::Timestamp;
+
+pub fn timestamp() -> Timestamp {
+	unimplemented!()
+}
+
+pub fn set_timestamp(_now: Timestamp) {
+	unimplemented!()
+}
